devices: debian: refactor ISC provisioner into separate class
